requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
feedgenerator==2.1.0
tweepy==4.14.0
facebook-sdk==3.1.0
python-linkedin==4.1
Mastodon.py==1.8.1
discord-webhook==1.3.0
PyYAML==6.0.1
Jinja2==3.1.2
markdown==3.5.1
GitPython==3.1.40
PyGithub==2.1.1
python-gitlab==3.15.0
bitbucket-api==0.5.0
requests==2.31.0
sqlalchemy==2.0.23
difflib-sequences==1.1.1
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
